Add normalization but lesser good result #dafuq

+++ alignment/src/main/scala/com/github/cowa/helpers/CosineSimilarity.scala
@@ -3,18 +3,18 @@ package com.github.cowa.helpers
 import language.postfixOps
 
 object CosineSimilarity {
-  def apply(t1: Map[String, Int], t2: Map[String, Int]): Double = {
-    val m = scala.collection.mutable.HashMap[String, (Int, Int)]() // word, t1 freq, t2 freq
+  def apply(t1: Map[String, Double], t2: Map[String, Double]): Double = {
+    val m = scala.collection.mutable.HashMap[String, (Double, Double)]() // word, t1 freq, t2 freq
 
     val sum1 = t1.foldLeft(0d) { case (sum, (word, freq)) =>
-      m += word ->(freq, 0)
+      m += word ->(freq, 0d)
       sum + freq
     }
 
     val sum2 = t2.foldLeft(0d) { case (sum, (word, freq)) =>
       m.get(word) match {
         case Some((freq1, _)) => m += word -> (freq1, freq)
-        case None => m += word ->(0, freq)
+        case None => m += word ->(0d, freq)
       }
       sum + freq
     }
